% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_email}
\alias{add_email}
\title{Adds email icon & link (to be used, eg in the 'footer-social' container)}
\usage{
add_email(
  email = "g.baio@ucl.ac.uk",
  title = "Email me",
  fill = "#bcc0c4",
  scale = "0.8em",
  style = NULL
)
}
\arguments{
\item{email}{The email address}

\item{title}{The alt text to be shown upon hovering the icon (default at
'Email me')}

\item{fill}{A string or a HEX code with the color of the icon}

\item{scale}{A number indicating the scale of the icon (default = 0.8)}

\item{style}{A string with some css style options (e.g. 'border-bottom: 0px;')}
}
\description{
Adds email icon & link (to be used, eg in the 'footer-social' container)
}
