% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_github}
\alias{add_github}
\title{Adds github icon & link (to be used, eg in the 'footer-social' container)}
\usage{
add_github(
  url = "https://github.com/giabaio",
  title = "Check out my repos",
  fill = "#bcc0c4",
  scale = 0.8,
  style = NULL
)
}
\arguments{
\item{url}{The URL of the website}

\item{title}{The alt text to be shown upon hovering the icon (default at
'Check out my repos')}

\item{fill}{A string or a HEX code with the color of the icon}

\item{scale}{A number indicating the scale of the icon (default = 0.8)}

\item{style}{A string with some css style options (e.g. 'border-bottom: 0px;')}
}
\description{
Adds github icon & link (to be used, eg in the 'footer-social' container)
}
