% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_instagram}
\alias{add_instagram}
\title{Adds instagram icon & link (to be used, eg in the 'footer-social' container)}
\usage{
add_instagram(
  url = "https://www.instagram.com/ucl.stats/",
  title = "Follow us on Instagram",
  fill = "#2867b2",
  scale = 0.8
)
}
\arguments{
\item{url}{The URL of the website}

\item{title}{The alt text to be shown upon hovering the icon (default at
'Follow us on Instagram')}

\item{fill}{A string or a HEX code with the color of the icon}

\item{scale}{A number indicating the scale of the icon (default = 0.8)}
}
\description{
Adds instagram icon & link (to be used, eg in the 'footer-social' container)
}
