% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_linkedin}
\alias{add_linkedin}
\title{Adds linkedin icon & link (to be used, eg in the 'footer-social' container)}
\usage{
add_linkedin(
  url = "https://www.linkedin.com/in/gianluca-baio-b893879/",
  title = "Follow me on LinkedIn",
  fill = "#2867b2",
  scale = 0.8,
  style = NULL
)
}
\arguments{
\item{url}{The URL of the website}

\item{title}{The alt text to be shown upon hovering the icon (default at
'Follow me on LinkedIn')}

\item{fill}{A string or a HEX code with the color of the icon}

\item{scale}{A number indicating the scale of the icon (default = 0.8)}

\item{style}{A string with some css style options (e.g. 'border-bottom: 0px;')}
}
\description{
Adds linkedin icon & link (to be used, eg in the 'footer-social' container)
}
