% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_mastodon}
\alias{add_mastodon}
\title{Adds mastodon icon & link (to be used, eg in the 'footer-social' container)}
\usage{
add_mastodon(
  url = "https://mas.to/@gianlubaio",
  title = "Follow me on Mastodon",
  fill = "#bcc0c4",
  scale = ".8em",
  style = NULL
)
}
\arguments{
\item{url}{The URL of the Twitter account}

\item{title}{The alt text to be shown upon hovering the icon (default at
'Follow me on Mastodon')}

\item{fill}{A string or a HEX code with the color of the icon}

\item{scale}{A number indicating the scale of the icon (default = 0.8)}

\item{style}{A string with some css style options (e.g. 'border-bottom: 0px;')}
}
\description{
Adds mastodon icon & link (to be used, eg in the 'footer-social' container)
}
