% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_podcast}
\alias{add_podcast}
\title{Adds podcast icon & link (to be used, eg in the 'footer-social' container)}
\usage{
add_podcast(
  url = "https://soundcloud.com/uclsound/sets/sample-space",
  title = "Random Talks",
  fill = "#FE5000",
  scale = 0.8,
  style = NULL
)
}
\arguments{
\item{url}{The URL of the website}

\item{title}{The alt text to be shown upon hovering the icon (default at
'Random Talks')}

\item{fill}{A string or a HEX code with the color of the icon}

\item{scale}{A number indicating the scale of the icon (default = 0.8)}

\item{style}{A string with some css style options (e.g. 'border-bottom: 0px;')}
}
\description{
Adds podcast icon & link (to be used, eg in the 'footer-social' container)
}
