% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_social}
\alias{add_social}
\title{Adds social media icon & link (to be used, eg in the 'footer-social' container)}
\usage{
add_social(
  social,
  url,
  title,
  fill = "#bcc0c4",
  scale = ".8em",
  style = "border-bottom: 0px!important;"
)
}
\arguments{
\item{social}{The name of the social media icon to add (a 'fontawesome' icon)}

\item{url}{The URL of the Twitter account}

\item{title}{The alt text to be shown upon hovering the icon (default at
'Follow me on Twitter')}

\item{fill}{A string or a HEX code with the color of the icon}

\item{scale}{A number indicating the scale of the icon (default = 0.8)}

\item{style}{A string with some css style options (e.g. 'border-bottom: 0px;')}
}
\description{
Adds social media icon & link (to be used, eg in the 'footer-social' container)
}
