% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_twitter}
\alias{add_twitter}
\title{Adds twitter icon & link (to be used, eg in the 'footer-social' container)}
\usage{
add_twitter(
  url = "https://twitter.com/giabaio",
  title = "Follow me on Twitter",
  fill = "#bcc0c4",
  scale = 0.8
)
}
\arguments{
\item{url}{The URL of the Twitter account}

\item{title}{The alt text to be shown upon hovering the icon (default at
'Follow me on Twitter')}

\item{fill}{A string or a HEX code with the color of the icon}

\item{scale}{A number indicating the scale of the icon (default = 0.8)}
}
\description{
Adds twitter icon & link (to be used, eg in the 'footer-social' container)
}
