% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_website}
\alias{add_website}
\title{Adds website icon & link (to be used, eg in the 'footer-social' container)}
\usage{
add_website(
  url = "https://gianluca.statistica.it",
  title = "Visit my website",
  fill = "#bcc0c4",
  scale = ".8em",
  style = NULL
)
}
\arguments{
\item{url}{The URL of the website}

\item{title}{The alt text to be shown upon hovering the icon (default at
'Visit my website')}

\item{fill}{A string or a HEX code with the color of the icon}

\item{scale}{A number indicating the scale of the icon (default = 0.8)}

\item{style}{A string with some css style options (e.g. 'border-bottom: 0px;')}
}
\description{
Adds website icon & link (to be used, eg in the 'footer-social' container)
}
