% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{include_fig}
\alias{include_fig}
\title{Macro to include figures from file}
\usage{
include_fig(img, dir = "./img", width = "100\%", title = "INCLUDE TEXT HERE")
}
\arguments{
\item{img}{The string with the name of the file with the image to be included}

\item{dir}{The path to the image. By default, assumes that the images to be
loaded are in the \code{img} folder under the current directory, ie where the
current Rmd presentation is stored. But a full directory can be passed on}

\item{width}{The width of the image (default 100\%)}

\item{title}{The string of text to define the alt tag for the image}
}
\description{
Macro to include figures from file
}
