% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{logo_stats}
\alias{logo_stats}
\title{Creates html code to include the UCL Stats logo on the title slide of a
quarto presentation. Needs to do this because quarto messes up with the
path of the main theme css, so cannot include the logo using the css code}
\usage{
logo_stats(url = "assets/images/UCL_Stats_logo.jpeg")
}
\arguments{
\item{url}{The path to the file}
}
\description{
Creates html code to include the UCL Stats logo on the title slide of a
quarto presentation. Needs to do this because quarto messes up with the
path of the main theme css, so cannot include the logo using the css code
}
