% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{postit}
\alias{postit}
\title{Makes css/html code to create post-its with variable inputs}
\usage{
postit(
  text = "This is some text",
  top = "50\%",
  left = "2.5\%",
  fontsize = "85\%",
  height = "4em",
  width = "4em",
  rotate = "6deg"
)
}
\arguments{
\item{text}{The text to be printed}

\item{top}{The position from the top of the slide (default "50\%")}

\item{left}{The position from the left of the slide (default "2.5\%")}

\item{fontsize}{The font size for the text (default as "85\%" of normal size)}

\item{height}{The height of the enclosing post it (default at "4em")}

\item{width}{The width of the enclosing post it (default at "4em")}

\item{rotate}{The rotation for the post it (default at "6deg")}
}
\description{
Makes css/html code to create post-its with variable inputs
}
