% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{publish_slides}
\alias{publish_slides}
\title{Publish the slides onto a specific folder}
\usage{
publish_slides(target)
}
\arguments{
\item{target}{The path to the folder in which the relevant files and folders
should be copied.}
}
\description{
Copies the relevant files/folders onto a specific target, which is used to
publish the slides online
}
\examples{
\dontrun{
publish_slides("~/Desktop/test")
}

}
\author{
Gianluca Baio
}
