% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{revealjs2pdf}
\alias{revealjs2pdf}
\title{Prints the revealjs slides to pdf using decktape}
\usage{
revealjs2pdf(
  filein,
  fileout,
  size = "1600x900",
  pause = "3000",
  fragment = TRUE,
  range,
  compression = "ebook"
)
}
\arguments{
\item{filein}{The path to the html revealjs slides file}

\item{fileout}{The path to the pdf file to be created}

\item{size}{The size in pixels of the slides (default "1600x900")}

\item{pause}{The pause duration before the next slide (longer time may be
needed for the page to fully load to capture the whole content). Default to
"3000"}

\item{fragment}{Should fragments (frame animation) all be printed. Default
to TRUE (in which case there's a single slide per animation)}

\item{range}{The range of pages to be printed (e.g. "1-3,5-8")}

\item{compression}{The compression level for the resulting pdf file. Default
to "ebook" and other options are "screen" (smallest file/worst resolution),
"printer" (bigger than "ebook", but larger), "prepress" (best resolution but
larger file). This parameter is passed to ps2pdf}
}
\description{
Prints the revealjs slides to pdf using `decktape` and `ps2pdf` for file
compression (both of which must be installed on the machine)
}
\examples{
\dontrun{
revealjs2pdf("index.html","slides.pdf")
}

}
\author{
Gianluca Baio
}
