% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{samptux}
\alias{samptux}
\title{Creates HTML code to include the samptux icon + link to gianluca.statistica.it}
\usage{
samptux(path = "assets/logo.png", width = "2.0\%", text = "")
}
\arguments{
\item{path}{The path to the image file (default "assets/logo.png")}

\item{width}{The width of the image (default "2.0\%")}

\item{text}{Some text to be added after the icon (default "")}
}
\description{
Creates HTML code to include the samptux icon + link to gianluca.statistica.it
}
