% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sftext}
\alias{sftext}
\title{Defines a LaTeX macro to typeset text using the default font
(as for the rest of the text in the slides)}
\usage{
sftext(x)
}
\arguments{
\item{x}{The text of string to be inserted in the LaTeX equation. Can also
use xaringan colors for equations, eg '.red[$$x^2$$]'. If a color is not
named (but defined in the .css file), need to use the MathJax command
\code{\\class}, eg \code{$\\class{color}{x^2}$}}
}
\description{
Defines a LaTeX macro to typeset text using the default font
(as for the rest of the text in the slides)
}
