% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides-package.R
\docType{package}
\name{slides-package}
\alias{slides}
\alias{slides-package}
\title{slides: Creates a set of templates for 'quarto' and 'xaringan' slides}
\description{
The package creates a set of templates for 'quarto' and 'xaringan' slides to be used for academic presentations.
}
\author{
\strong{Maintainer}: Gianluca Baio \email{g.baio@ucl.ac.uk} (\href{https://orcid.org/0000-0003-4314-2570}{ORCID})

Other contributors:
\itemize{
  \item Cara Thompson [contributor]
}

}
\keyword{internal}
