% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tab}
\alias{tab}
\title{Creates a 'tab' function that adds horizontal space(s)}
\usage{
tab(x = 1)
}
\arguments{
\item{x}{The number of 'tabs' to be added (defaults to 1)}
}
\description{
Creates a 'tab' function that adds horizontal space(s)
}
