% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{vspace}
\alias{vspace}
\title{Uses R to create HTML code to make space}
\usage{
vspace(space = "-20px")
}
\arguments{
\item{space}{The amount of vertical space to include (default to "-20px")}
}
\description{
Uses R to create HTML code to make space
}
